+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - plotBaseTiles function
 - Tile class with helper methods
 - Pan viewport helper
+- Coords and text drawing